pyproject.toml
# https://python-poetry.org/docs/libraries/#versioning
# https://pypi.org/project/hcaptcha-challenger/#history

[tool.poetry]
name = "hcaptcha-challenger"
version = "0.8.6"
description = "🥂 Gracefully face hCaptcha challenge with MoE(ONNX) embedded solution."
license = "GPL-3.0-or-later"
authors = ["QIN2DIM <yaoqinse@gmail.com>", "Bingjie Yan <bj.yan.pa@qq.com>"]
readme = "README.md"
homepage = "https://github.com/QIN2DIM/hcaptcha-challenger"
repository = "https://github.com/QIN2DIM/hcaptcha-challenger"
documentation = "https://github.com/QIN2DIM/hcaptcha-challenger"
keywords = ["hcaptcha", "hcaptcha-challenger", "hcaptcha-solver"]
packages = [{ include = "hcaptcha_challenger", format = "sdist" }]
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]

[virtualenvs]
# https://python-poetry.org/docs/configuration/#virtualenvsin-project
# Create a `.venv` virtual environment in the project root
in-project = true

# https://python-poetry.org/docs/pyproject/#dependencies-and-dependency-groups
[tool.poetry.dependencies]
python = "^3.8"

# Run `poetry install` in the project root
opencv-python = "^4.8.0.76"
numpy = "^1.24.4"
httpx = { version = "*", extras = ["http2"] }
onnxruntime = "^1.16.0"
loguru = "*"
tqdm = "*"
pyyaml = "*"
scikit-learn = "^1.3.1"

# The following dependencies will not be installed
# when you just run `poetry install` or `pip install hcaptcha-challenger`
undetected-chromedriver = { version = "^3.5.3", optional = true }
webdriver-manager = { version = "^4.0.1", optional = true }
selenium = { version = "*", optional = true }
PyGithub = { version = "^1.59.1", optional = true }
istockphoto = { version = "0.1.2", optional = true }
playwright = "^1.38.0"
pyquery = "^2.0.0"
pika = "^1.3.2"
psycopg2-binary = "^2.9.9"
retry = "^0.9.2"
asyncpg = "^0.28.0"

[tool.poetry.group.test.dependencies]
# https://docs.pytest.org/en/stable/reference/plugin_list.html#plugin-list
# https://docs.pytest.org/en/stable/contents.html
pytest = "*"
# https://pytest-asyncio.readthedocs.io/en/latest/
pytest-asyncio = "*"
# https://pytest-cov.readthedocs.io/en/latest/readme.html
pytest-cov = "*"
# https://pytest-xdist.readthedocs.io/en/stable/
pytest-xdist = { version = "*", extras = ["psutil"] }
# https://black.readthedocs.io/en/stable/
black = { version = "*", extras = ["jupyter"] }
# https://github.com/pythonprofilers/memory_profiler
memory-profiler = "*"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/reference/reference.html#configuration-options
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"

[tool.poetry.extras]
# 1. Running `poetry install -E [extras]` will install the general dependencies
# and specified dependencies at the same time.

# 2. The `-E` parameter can accept multiple extras-names,
# that is, if you only specify one extra, poetry will first uninstall the other extras installed
# and then try to install the extra you specified

# 3. Install all extra dependencies with `--all-extras`.

# Developer: `poetry install -E selenium`
# User: `pip install hcaptcha-challenger[selenium]`
selenium = ["selenium", "undetected-chromedriver", "webdriver-manager"]

# Developer: `poetry install -E playwright`
# User: `pip install hcaptcha-challenger[playwright]`
playwright = ["playwright"]

# Developer: `poetry install -E sentinel`
# --> ci: sentinel
# --> ci: collector
sentinel = ["playwright", "PyGithub"]

# Developer: `poetry install -E istock`
# --> pull open source datasets
istock = ["istockphoto"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
